requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
rembg
onnxruntime
opencv-python
numpy
requests
pydantic

